refactor out of speedups & remove duplicate funcs
